move configuration to its own module
